main markup in proccess
