+++ src/index.js
@@ -1,4 +1,2 @@
 import './js/api';
 import './js/base';
-import './js/main-test';
-import './js/pagination';
