+++ src/partials/header.html
@@ -1,42 +1,45 @@
-
-
-
 <header class="header">
-    <div class="logo">
-       
+  <div class="logo">
     <a href="">
-        <img srcset="./images/logo_mob@1x.png 1x, 
-        ./images/logo_mob@2x.png 2x" 
-        src="./images/logo_mob@1x.png" alt="">
+      <img
+        srcset="./images/logo_mob@1x.png 1x, ./images/logo_mob@2x.png 2x"
+        src="./images/logo_mob@1x.png"
+        alt=""
+      />
     </a>
+  </div>
 
+  <div class="header-overlay"></div>
+  <div class="container">
+    <div class="header-box">
+      <h1 class="header-title">FIND BEST EVENTS AROUND THE WORLD</h1>
+
+      <div class="post">
+        <form class="form">
+          <input
+            type="text"
+            class="input input-text"
+            placeholder="Start searching"
+          />
+          <button type="submit" class="button">
+            <svg class="btn-favor-icon" width="12" height="12">
+              <use
+                class="header-icon"
+                href="./images/search_svg.svg#icon-search_svg-1"
+              ></use>
+            </svg>
+          </button>
+        </form>
+        <form class="form">
+          <select
+            required
+            class="input select-country"
+            placeholder="Start searching"
+          >
+            <option value="" class="country-option"></option>
+          </select>
+        </form>
+      </div>
     </div>
-
-
-    
-    
-    <div class="header-overlay"></div>
-    <div class="container">
-        <div class="header-box">
-            <h1 class="header-title">FIND BEST EVENTS AROUND THE WORLD</h1>
-    
-            <div class="post">
-                <form class="form">
-                    <input type="text" class="input input-text" placeholder="Start searching" />
-                    <button type="submit" class="button">
-                        <svg class="btn-favor-icon" width="12" height="12">
-                            <use class="header-icon" href="./images/search_svg.svg#icon-search_svg-1"></use>
-                        </svg>
-                    </button>
-                </form>
-                <form class="form">
-                    <select required class="input select-country" placeholder="Start searching">
-                        <option value="" class="country-option"></option>
-                       
-                    </select>
-                    
-                </form>
-            </div>
-        </div>
-    </div>
+  </div>
 </header>
