+++ src/partials/main.html
@@ -1,4 +1,4 @@
-<main>
+<main class="main">
   <section class="section">
     <div class="container">
       <ul class="gallery">
