+++ src/sass/layout/_main.scss
@@ -1,3 +1,7 @@
+.main {
+  margin-top: -102px;
+  padding-top: 26px;
+}
 .gallery {
   display: flex;
   flex-wrap: wrap;
@@ -9,7 +13,7 @@
     column-gap: 32px;
   }
 
-  padding-top: 30px;
+  // padding-top: 30px;
 }
 
 .gallery__itams {
@@ -26,6 +30,8 @@
 }
 
 .gallery-link__img {
+  position: relative;
+  z-index: 1;
   border-radius: 50px 0px;
   width: calc(100% - 10px);
   @include mq(mob-only) {
@@ -57,7 +63,6 @@
   border-radius: 50px 0px;
   width: calc(100% - 10px);
   height: 100%;
-  z-index: -1;
 
   @include mq(tab) {
     width: 141px;
